moved contact to lib
