+++ projects/hpfb/sdk/ui/contact/contact.details/contact.details.component.ts
