+++ projects/hpfb/sdk/ui/contact/contact.list/contact-list.service.ts
@@ -1,10 +1,10 @@
 import {Injectable} from '@angular/core';
 import {FormArray, FormBuilder, FormGroup} from '@angular/forms';
 // import {CompanyContactRecordService} from '../company-contact-record/company-contact-record.service';
-import {IMasterDetails} from '../../master-details';
-
+import { IMasterDetails } from '../../list/master-details';
 // import {ContactDetailsService} from '../contact.details/contact.details.service';
-import {ListService} from '../../list-service';
+import {ListService} from '../../list/list-service';
+
 
 @Injectable()
 export class ContactListService extends ListService implements IMasterDetails {
