+++ projects/hpfb/sdk/ui/list/list-service.ts
@@ -1,6 +1,3 @@
-import {FormArray, FormGroup} from '@angular/forms';
-
-
 export abstract class ListService {
 
   /**
