+++ projects/hpfb/sdk/ui/list/master-details.ts
@@ -1,5 +1,7 @@
 import { FormGroup } from '@angular/forms';
 
+// todo rename the class to ??
+
 export interface IMasterDetails {
   deleteModelRecord(id: number);
   getModelRecordList();
