+++ projects/hpfb/sdk/ui/public-api.ts
@@ -1,17 +1,30 @@
+export * from './layout/layout.component';
+
 export * from './expander/expander.component';
+
 export * from './error-msg/control-messages/control-messages.component';
 export * from './error-msg/error-summary/error-summary.component';
+
 export * from './address/address.details/address.details.component';
+
+export * from './contact/contact.details/contact.details.component';
+export * from './contact/company-contact-record/company-contact-record.component';
+export * from './contact/contact.list/contact.list.component';
+export * from './contact/contact.list/contact-list.service';
+
 export * from './file-io/greeter/greeter.component';
-export * from './layout/layout.component';
 export * from './file-io/convert-results'
 export * from './file-io/filereader/filereader.component';
 export * from './file-io/file-io.module';
+
 export * from './common.constants';
+
 export * from './utils/utils.service';
 export * from './validation/validation.service';
+
 export * from './pipes/form-control.pipe';
 export * from './pipes/json-keys.pipe';
+
 export * from './common.form.dendency.module';
 export * from './ui.module';
