+++ projects/hpfb/sdk/ui/ui.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CommonFormDendencyModule } from './common.form.dendency.module';
 import { FileIoModule } from './file-io/file-io.module';
+
 import { LayoutComponent } from './layout/layout.component';
 import { ExpanderComponent } from './expander/expander.component';
 import { ControlMessagesComponent } from './error-msg/control-messages/control-messages.component';
@@ -12,6 +13,10 @@ import { AddressDetailsService } from './address/address.details/address.details
 import { UtilsService } from './utils/utils.service';
 import { FormControlPipe } from './pipes/form-control.pipe';
 import { JsonKeysPipe } from './pipes/json-keys.pipe';
+import { CompanyContactRecordComponent } from './contact/company-contact-record/company-contact-record.component';
+import { ContactDetailsComponent } from './contact/contact.details/contact.details.component';
+import { ContactListComponent } from './contact/contact.list/contact.list.component';
+
 
 @NgModule({
   declarations: [
@@ -20,11 +25,18 @@ import { JsonKeysPipe } from './pipes/json-keys.pipe';
     ControlMessagesComponent,
     ErrorSummaryComponent,
     AddressDetailsComponent,
+    CompanyContactRecordComponent,
+    ContactDetailsComponent,
+    ContactListComponent,
     GreeterComponent,
     FormControlPipe,
     JsonKeysPipe,
   ],
-  imports: [CommonModule, CommonFormDendencyModule, FileIoModule],
+  imports: [
+    CommonModule,
+    CommonFormDendencyModule,
+    FileIoModule
+  ],
   providers: [AddressDetailsService, UtilsService],
   exports: [
     LayoutComponent,
@@ -32,6 +44,9 @@ import { JsonKeysPipe } from './pipes/json-keys.pipe';
     ControlMessagesComponent,
     ErrorSummaryComponent,
     AddressDetailsComponent,
+    CompanyContactRecordComponent,
+    ContactDetailsComponent,
+    ContactListComponent,
     GreeterComponent,
     FileIoModule,
     FormControlPipe,
@@ -39,3 +54,4 @@ import { JsonKeysPipe } from './pipes/json-keys.pipe';
   ],
 })
 export class UiModule {}
+
